Fixed segmentation fault that occurs in OPCODE on 64 bits systems.
Originally, OPCODE uses 32 bits integers to store pointer values. This is
a non-issue on 32 bits systems, but on 64 machines this of course causes
a nice segfault.

Changed the type used by OPCODE to store "pointer integers" from udword to
size_t. In pratice this means that 32 bits systems will use 32 bits
integers, and 64 bits systems will use 64 bits integers.
It has been suggested to use uintptr_t instead, but this type only appears
in C99 and not C++98.